Add rich info to dashboard labels
