+++ BTCPayServer/Components/StoreRecentTransactions/Default.cshtml
@@ -51,15 +51,26 @@
                         </a>
                     </td>
                     <td>
-                        <div class="d-flex flex-wrap gap-2 align-items-center">
-                            @foreach (var label in tx.Labels)
-                            {
-                                <a asp-controller="UIWallets" asp-action="WalletTransactions" asp-route-labelFilter="@label.Key" asp-route-walletId="@Model.WalletId"
-                                   class="transaction-label" style="--label-bg:@label.Value;--label-fg:@ColorPalette.Default.TextColor(label.Value)">
-                                    @label.Key
-                                </a>
-                            }
-                        </div>
+                        @if (tx.Labels.Any())
+                        {
+                            <div class="d-flex flex-wrap gap-2 align-items-center">
+                                @foreach (var label in tx.Labels)
+                                {
+                                    <div class="transaction-label" style="--label-bg:@label.Color;--label-fg:@label.TextColor">
+                                        <span>@label.Text</span>
+                                        @if (!string.IsNullOrEmpty(label.Link))
+                                        {
+                                            <a class="transaction-label-info transaction-details-icon" href="@label.Link"
+                                               target="_blank" rel="noreferrer noopener" title="@label.Tooltip" data-bs-html="true"
+                                               data-bs-toggle="tooltip" data-bs-custom-class="transaction-label-tooltip">
+                                                <vc:icon symbol="info" />
+                                            </a>
+                                        }
+                                    </div>
+                                }
+                            </div>
+                        }
+                        
                     </td>
                     @if (tx.Positive)
                     {
