+++ BTCPayServer/Services/Labels/Label.cs
@@ -1,8 +1,6 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using BTCPayServer.Client.Models;
-using NBitcoin;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
